+++ shopsystem-service/src/main/java/com/yandex/market/shopservice/controllers/branch/BranchController.java
@@ -19,5 +19,6 @@ public class BranchController {
     @ResponseStatus(HttpStatus.CREATED)
     public void createNewBranch(@RequestBody @Valid BranchDto branchDto) {
         log.info("Received a request to create new branch for shop system: %s".formatted(branchDto));
+        branchService.createBranch(branchDto);
     }
 }
