+++ shopsystem-service/src/main/java/com/yandex/market/shopservice/service/ExceptionResponse.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.http.HttpStatusCode;
 
 import java.time.LocalDateTime;
 import java.util.List;
